change key representation
Change the order in the key representation, so that artist is first,
album is second and title is last. This should provide better results
when the dictionary is sorted (recordings from the same album by the
same artist will be grouped together).

This is a breaking change: existing cache items will not be recognized
anymore because of key difference.