repeatedly ask for user selection
Don't exit when user provides an out-of-bounds recording index when
presented with multiple search results. In that case repeat the question
until a valid answer is provided. Allow the user to discard the search
result by selecting 0.

Allow get_recording_mbid() and select_recording() to return None, and
handle that case instead of raising and catching exceptions.

Change remaining double quotes to single quotes (an exception is the
recording title in search results, which is still printed in double
quotes).

Remove commented out code.