implement MusicBrainz release cache
Implement a cache for storing MusicBrainz release information. This is
based on RecordingCache object, but a bit more complicated. There's room
for improvement.

Change mb_release_lookup.py to retrieve more information about the
release, and to use the release cache. For now it only stores retrieved
release data in the cache without checking if the release data was
previously stored. Create and use main().

During testing it became clear that relying on __del__() to correctly
save cache state is a bad idea. A better way would be to provide a
dedicated method for saving state, and implementing __entry__() and
__exit__() methods for context management. That way using cache would
be similar to using regular file objects, where close() method can be
either called directly by the user, or can be called automatically when
with statement is used.