store last update and lookup times in cache
Change cache structure so that information about last update time and
last lookup time is stored for each index entry.

This complicates the code, but it may be useful to invalidate cache
entries if the cache grows too large.

This is a breaking change of cache format: existing caches become
invalid.