don't call ReleaseCache destructor if constructor fails
Up to now the destructor was always called, even when the cache index
was not loaded (e.g. because it was not valid JSON). In that case the
cache was empty, and remove_orphans() was removing all cache content.

Call the destructor only if the cache was created (i.e. the index file
was successfully loaded).