implement inter-process cache locking mechanism
Implement an exclusive, file-based lock to prevent concurrent access by
multiple processes. While an exclusive lock with no regard of the access
type may be suboptimal, it's not possible to predict if read-only access
will be sufficient, or if write access will be required as well.

The lock is acquired for the entire lifetime of the cache object. This
may block for too long, but the implementation does not easily allow
more fine-grained access control.

Due to fcntl usage, the implementation works only on POSIX-compliant
operating systems. Furthermore, it does not prevent concurrent access
by multiple threads in the same process. Exclusive access is ensured
only for different processes.