introduce low-level cache inheritance
Create a base low-level cache class. Refactor entity-specific low-level
caches by moving __init__(), __del__() and __str__() to the base class.
Rename __str__() to __repr__() because it provides a low-level JSON data
dump.

Unify cache initialization logic. The recording cache is now initially
None, and is initialized to an empty dictionary if the cache does not
exist yet. This aligns recording and release cache implementations and
allows moving cache initialization logic to the base class.

Change the order of operations in _ReleaseCache.__del__(): orphaned
release files are removed first, and the cache index is updated last.
super().__del__() must be called last, because it releases the cache
lock. The change of the order should make no difference in practice.