pass all entity parameters as a dictionary
Different caches require different parameters to their methods. This
makes it difficult to refactor common parts to the base class, because
the number of parameters can be different depending on entity type.

One way to work around this is to pass all entity parameters in a dict.
That way all methods can receive a single parameter, and extract needed
fields from it.