implement dedicated parameter objects
Using dictionaries as parameter sets was a quick-and-dirty proof of
concept. Now that it's clear that this approach to passing the entity
parameters makes sense, create dedicated data classes for storing them.
Adapt both the high-level and low-level cache classes to use the new
parameter objects.

An additional benefit is that it is now possible to move cache key
creation as a method to the entity-specific parameter class. Static
methods defined in low-level cache classes are no longer needed.

Configure MyPy. Because the new parameter objects are implemented as
data classes, type checking becomes important.