implement copy_recordings
Add program for copying recordings from a release to the recording
cache. Install the program as a console script.

This program is directly using protected attributes of _Cache, but this
is still internal usage in scope of one package. I think it's fine, so
the issue reported by Pylint is silenced.

MyPy reports an issue with remove_featured() function when the second
argument is None, but this is actually a problem in mbnames. While the
function correctly handles None parameter, its type hint doesn't specify
that this parameter is optional. MyPy issue is silenced until it's fixed
in mbnames.